Add a comment system on pelican.
This system is based on email.
Users have to send comment by mail :
- By thereselve using right format (pretty difficult)
- Using a html form "actioning" to a script sending the mail as necessary
  (far easier for them)

You will have to setup the following config variable :
- IMAP_HOSTNAME : The hostname where to read comments/emails.
- IMAP_USERNAME : The username to use.
- IMAP_PASSWORD : The password to use.
- IMAP_READBOX  : In wish directory read emails (INBOX, INBOX/validated)

In samples/commentsystem :
- send_comment.py :
  A python cgi script to send the mail where it need to be sent
- comments.html
  A template to insert at the end of your article template:

    <div id="article">
      ...
      {{ article.content }}
      ...
      {% include 'comments.html' %}
      ...
    </div>

This code worked on my early tests. However it never has been push
in production. It is published here cause somes want it to be shared.
Feel free to use it as you need it but use it at your own risk.

License is the same of pelican (AGPL) cause code is part of pelican.
But I will not bother you if you take it in your project on
another license. (Maybe write my name somewhere on your sources :P )